+++ src/Utility/Fs/Enum/FinderMode.php
@@ -14,7 +14,7 @@ declare(strict_types=1);
  * @since         5.3.0
  * @license       https://opensource.org/licenses/mit-license.php MIT License
  */
-namespace Cake\Filesystem\Enum;
+namespace Cake\Utility\Fs\Enum;
 
 /**
  * Enum for Finder iteration modes
